Update Frame.java
This is the pull request as Samuel requested. I added the clone() method and I also added a long timeStamp for future implementation, because I think it's better that the frame itself contains the time signature when it was shot rather than reading the grabbers's timestamp.

Update Frame.java

This is the pull request as Samuel requested. I added the clone() method and I also added a long timeStamp for future implementation, because I think it's better that the frame itself contains the time signature when it was shot rather than reading the grabbers's timestamp.

Update Frame.java

Update Frame.java

Update Frame.java